Share router logic
